+++ dev/user.clj
@@ -1,0 +1,2 @@
+(ns user
+  (:require [clojure.tools.namespace.repl :as repl]))