+++ test/selmer/benchmark.clj
@@ -3,10 +3,11 @@
             [selmer.parser :refer :all]
             [selmer.filters :as filters]
             [selmer.util :refer :all]
-            [criterium.core :as criterium])
-  (:import java.io.StringReader))
+            [criterium.core :as criterium]
+            [clojure.string :as string]
+            [selmer.parser :as parser]))
 
-(def user (repeat 10 [{:name "test" }]))
+(def user (repeat 10 [{:name "test"}]))
 
 (def nested-for-context {:users (repeat 10 user)})
 
@@ -48,3 +49,25 @@
   (filters/add-filter! :inc (fn [^String s] (str (inc (Integer/parseInt s)))))
   (criterium/quick-bench
     (render (str "{{bar" filter-chain "}}") {:bar "0"})))
+
+(deftest ^:benchmark if-bench
+  (println "BENCH: Many . acceses in an if clause")
+  (criterium/quick-bench
+    (render (string/join "" (repeat 1000 "{% if p.a.a.a.a %}{% endif %}"))
+            {})))
+
+(deftest ^:benchmark many-numeric-if-clauses-bench
+  (println "BENCH: for loop with a numeric if clause in it")
+  (reset! parser/templates {})
+  (cache-on!)
+  (render-file "templates/numerics.html" {:ps []})
+  (criterium/quick-bench
+    (render-file "templates/numerics.html" {:ps (repeat 10000 "x")})))
+
+(deftest ^:benchmark many-any-if-clauses-bench
+  (println "BENCH: for loop with a if any clause in it")
+  (reset! parser/templates {})
+  (cache-on!)
+  (render-file "templates/any.html" {:products []})
+  (criterium/quick-bench
+    (render-file "templates/any.html" {:products (repeat 10000 {})})))